+++ lib/src/all_weapons.dart
@@ -7,6 +7,7 @@ import 'database/weapons/ax_20_blaster_cannon.dart' as _i;
 import 'database/weapons/boot_spikes.dart' as _i;
 import 'database/weapons/brawl.dart' as _i;
 import 'database/weapons/chewbaccas_bowcaster.dart' as _i;
+import 'database/weapons/close_quarters_combat.dart' as _i;
 import 'database/weapons/collapsible_tonfa.dart' as _i;
 import 'database/weapons/combat_expertise.dart' as _i;
 import 'database/weapons/commands/coordinated_bombardment.dart' as _i;
@@ -17,6 +18,7 @@ import 'database/weapons/commands/zx_flame_projector.dart' as _i;
 import 'database/weapons/dh_17_blaster_pistol.dart' as _i;
 import 'database/weapons/dual_wester_35_blaster_pistols.dart' as _i;
 import 'database/weapons/e_11_blaster_rifle.dart' as _i;
+import 'database/weapons/e_11_d_blaster_rifle.dart' as _i;
 import 'database/weapons/e_web_heavy_blaster.dart' as _i;
 import 'database/weapons/ec_17_hold_out_blaster.dart' as _i;
 import 'database/weapons/fence_cutting_blades.dart' as _i;
@@ -28,6 +30,7 @@ import 'database/weapons/frenzy.dart' as _i;
 import 'database/weapons/grappling_claws.dart' as _i;
 import 'database/weapons/hans_dl_44_blaster_pistol.dart' as _i;
 import 'database/weapons/kashyyyk_pistol.dart' as _i;
+import 'database/weapons/krennics_dt_29_blaster_pistol.dart' as _i;
 import 'database/weapons/lateral_quad_mk_2_e_w_cannons.dart' as _i;
 import 'database/weapons/leias_defender_sporting_blaster.dart' as _i;
 import 'database/weapons/lukes_dl_44_blaster_pistol.dart' as _i;
@@ -38,6 +41,7 @@ import 'database/weapons/officers_rk_3_blaster_pistol.dart' as _i;
 import 'database/weapons/overpower.dart' as _i;
 import 'database/weapons/relby-v-10-mortar-rifle.dart' as _i;
 import 'database/weapons/ryyk_blade.dart' as _i;
+import 'database/weapons/se_14_n_light_blaster.dart' as _i;
 import 'database/weapons/unarmed.dart' as _i;
 import 'database/weapons/upgrades/a_180_blaster_pistol.dart' as _i;
 import 'database/weapons/upgrades/a_180_blaster_rifle.dart' as _i;
@@ -46,8 +50,12 @@ import 'database/weapons/upgrades/a_300_short_range_config.dart' as _i;
 import 'database/weapons/upgrades/bistan.dart' as _i;
 import 'database/weapons/upgrades/bowcaster.dart' as _i;
 import 'database/weapons/upgrades/dh_447_sniper_rifle.dart' as _i;
+import 'database/weapons/upgrades/dlt_19_d_heavy_blaster_rifle.dart' as _i;
 import 'database/weapons/upgrades/dlt_19_heavy_blaster_rifle.dart' as _i;
 import 'database/weapons/upgrades/dlt_19x_sniper_rifle.dart' as _i;
+import 'database/weapons/upgrades/dt_f_16_blaster_rifle.dart' as _i;
+import 'database/weapons/upgrades/e_11_d_grenade_launcher.dart' as _i;
+import 'database/weapons/upgrades/e_11_focused_fire.dart' as _i;
 import 'database/weapons/upgrades/electrostaff.dart' as _i;
 import 'database/weapons/upgrades/flamethrower.dart' as _i;
 import 'database/weapons/upgrades/grenades/concussion_grenades.dart' as _i;
@@ -68,6 +76,7 @@ import 'database/weapons/upgrades/mpl_57_barrage_launcher.dart' as _i;
 import 'database/weapons/upgrades/mpl_57_ion_launcher.dart' as _i;
 import 'database/weapons/upgrades/pao.dart' as _i;
 import 'database/weapons/upgrades/proton_charge.dart' as _i;
+import 'database/weapons/upgrades/rt_97_c_rifle_pintle.dart' as _i;
 import 'database/weapons/upgrades/scatter_gun.dart' as _i;
 import 'database/weapons/upgrades/sonic_charge.dart' as _i;
 import 'database/weapons/upgrades/t_7_ion_disrupter_rifle.dart' as _i;
@@ -87,6 +96,7 @@ final aggregate = [
   _i.bootSpikes,
   _i.brawl,
   _i.chewbaccasBowcaster,
+  _i.closeQuartersCombat,
   _i.collapsibleTonfa,
   _i.combatExpertise,
   _i.coordinatedBombardment,
@@ -97,6 +107,7 @@ final aggregate = [
   _i.dh17BlasterPistol,
   _i.dualWestar35BlasterPistols,
   _i.e11BlasterRifle,
+  _i.e11DBlasterRifle,
   _i.eWebHeavyBlaster,
   _i.ec17HoldOutBlaster,
   _i.fenceCuttingBlades,
@@ -108,6 +119,7 @@ final aggregate = [
   _i.grapplingClaws,
   _i.hansDl44BlasterPistol,
   _i.kashyyykPistol,
+  _i.krennicsDt29BlasterPistol,
   _i.lateralQuadMK2EWCannons,
   _i.leiasDefenderSportingBlaster,
   _i.lukesDl44BlasterPistol,
@@ -118,6 +130,7 @@ final aggregate = [
   _i.overpower,
   _i.relbyV10MortarRifle,
   _i.ryykBlade,
+  _i.se14NLightBlaster,
   _i.unarmed,
   _i.a180BlasterPistol,
   _i.a180BlasterRifle,
@@ -126,8 +139,12 @@ final aggregate = [
   _i.bistan,
   _i.bowcaster,
   _i.dh447SniperRifle,
-  _i.dlt19HHeavyBlasterRifle,
+  _i.dlt19DHeavyBlasterRifle,
+  _i.dlt19HeavyBlasterRifle,
   _i.dlt19XSniperRifle,
+  _i.dtF16BlasterRifle,
+  _i.e11DGrenadeLauncher,
+  _i.e11DFocusedFire,
   _i.electrostaff,
   _i.flamethrower,
   _i.concussionGrenades,
@@ -146,6 +163,7 @@ final aggregate = [
   _i.mpl57IonLauncher,
   _i.pao,
   _i.protonCharge,
+  _i.rt97CBlasterRifle,
   _i.scatterGun,
   _i.sonicCharge,
   _i.t7IonDisrupterRifle,
@@ -167,6 +185,7 @@ class Weapons {
   static final bootSpikes = _i.bootSpikes;
   static final brawl = _i.brawl;
   static final chewbaccasBowcaster = _i.chewbaccasBowcaster;
+  static final closeQuartersCombat = _i.closeQuartersCombat;
   static final collapsibleTonfa = _i.collapsibleTonfa;
   static final combatExpertise = _i.combatExpertise;
   static final coordinatedBombardment = _i.coordinatedBombardment;
@@ -177,6 +196,7 @@ class Weapons {
   static final dh17BlasterPistol = _i.dh17BlasterPistol;
   static final dualWestar35BlasterPistols = _i.dualWestar35BlasterPistols;
   static final e11BlasterRifle = _i.e11BlasterRifle;
+  static final e11DBlasterRifle = _i.e11DBlasterRifle;
   static final eWebHeavyBlaster = _i.eWebHeavyBlaster;
   static final ec17HoldOutBlaster = _i.ec17HoldOutBlaster;
   static final fenceCuttingBlades = _i.fenceCuttingBlades;
@@ -188,6 +208,7 @@ class Weapons {
   static final grapplingClaws = _i.grapplingClaws;
   static final hansDl44BlasterPistol = _i.hansDl44BlasterPistol;
   static final kashyyykPistol = _i.kashyyykPistol;
+  static final krennicsDt29BlasterPistol = _i.krennicsDt29BlasterPistol;
   static final lateralQuadMK2EWCannons = _i.lateralQuadMK2EWCannons;
   static final leiasDefenderSportingBlaster = _i.leiasDefenderSportingBlaster;
   static final lukesDl44BlasterPistol = _i.lukesDl44BlasterPistol;
@@ -198,6 +219,7 @@ class Weapons {
   static final overpower = _i.overpower;
   static final relbyV10MortarRifle = _i.relbyV10MortarRifle;
   static final ryykBlade = _i.ryykBlade;
+  static final se14NLightBlaster = _i.se14NLightBlaster;
   static final unarmed = _i.unarmed;
   static final a180BlasterPistol = _i.a180BlasterPistol;
   static final a180BlasterRifle = _i.a180BlasterRifle;
@@ -206,8 +228,12 @@ class Weapons {
   static final bistan = _i.bistan;
   static final bowcaster = _i.bowcaster;
   static final dh447SniperRifle = _i.dh447SniperRifle;
-  static final dlt19HHeavyBlasterRifle = _i.dlt19HHeavyBlasterRifle;
+  static final dlt19DHeavyBlasterRifle = _i.dlt19DHeavyBlasterRifle;
+  static final dlt19HeavyBlasterRifle = _i.dlt19HeavyBlasterRifle;
   static final dlt19XSniperRifle = _i.dlt19XSniperRifle;
+  static final dtF16BlasterRifle = _i.dtF16BlasterRifle;
+  static final e11DGrenadeLauncher = _i.e11DGrenadeLauncher;
+  static final e11DFocusedFire = _i.e11DFocusedFire;
   static final electrostaff = _i.electrostaff;
   static final flamethrower = _i.flamethrower;
   static final concussionGrenades = _i.concussionGrenades;
@@ -226,6 +252,7 @@ class Weapons {
   static final mpl57IonLauncher = _i.mpl57IonLauncher;
   static final pao = _i.pao;
   static final protonCharge = _i.protonCharge;
+  static final rt97CBlasterRifle = _i.rt97CBlasterRifle;
   static final scatterGun = _i.scatterGun;
   static final sonicCharge = _i.sonicCharge;
   static final t7IonDisrupterRifle = _i.t7IonDisrupterRifle;