+++ lib/src/database/weapons/close_quarters_combat.dart
@@ -1,0 +1,8 @@
+import 'package:swlegion/swlegion.dart';
+
+final closeQuartersCombat = Weapon.melee(
+  name: 'Close Quarters Combat',
+  dice: {
+    AttackDice.red: 1,
+  },
+);