+++ lib/src/database/weapons/e_11_d_blaster_rifle.dart
@@ -1,0 +1,10 @@
+import 'package:swlegion/swlegion.dart';
+
+final e11DBlasterRifle = Weapon(
+  name: 'E-11D Blaster Rifle',
+  minRange: 1,
+  maxRange: 3,
+  dice: {
+    AttackDice.black: 1,
+  },
+);