+++ lib/src/database/weapons/se_14_n_light_blaster.dart
@@ -1,0 +1,10 @@
+import 'package:swlegion/swlegion.dart';
+
+final se14NLightBlaster = Weapon(
+  name: 'SE-14n Light Blaster',
+  minRange: 1,
+  maxRange: 2,
+  dice: {
+    AttackDice.white: 2,
+  },
+);