+++ lib/src/database/weapons/upgrades/e_11_d_grenade_launcher.dart
@@ -1,0 +1,13 @@
+import 'package:swlegion/swlegion.dart';
+
+final e11DGrenadeLauncher = Weapon(
+  name: 'E-11D Grenade Launcher',
+  minRange: 1,
+  maxRange: 2,
+  dice: {
+    AttackDice.red: 1,
+  },
+  keywords: {
+    WeaponKeyword.kBlast: '',
+  },
+);