+++ lib/src/database/weapons/upgrades/e_11_focused_fire.dart
@@ -1,0 +1,13 @@
+import 'package:swlegion/swlegion.dart';
+
+final e11DFocusedFire = Weapon(
+  name: 'E-11D Focused Fire',
+  minRange: 1,
+  maxRange: 4,
+  dice: {
+    AttackDice.black: 1,
+  },
+  keywords: {
+    WeaponKeyword.kSuppressive: '',
+  },
+);